+++ RELEASE-NOTES.md
@@ -1,4 +1,5 @@
 ## 4.0-M4
 
 * #1 OpenTelemetry bootstrap
+* #3 Injectable OTel Resource with Bootique application name #3